further improvements

+++ src/main/scala/core/lars/package.scala
@@ -1,12 +1,10 @@
 package core
 
-import engine.{Time, Stream}
-
 /**
   * Created by FM on 29.04.16.
   */
 package object lars {
-  type WindowSize = Int
+  type WindowSize = Long
 
   //  type WindowFunction = (WindowSize) => (Stream, Time) => Stream
 //  type WindowFunction = (Stream, Time) => Stream
